incorporate CheckBuilder to AkkaCheckSupport
